Rename "networkendpointgroup" to "neg"
